refactor: split config utility methods off into `ConfigUtil`
